Refactor: split api controller to user and table controller, rename the old table controller
