+++ src/main/java/com/coodecool/pickyourspot/controller/TableController.java
@@ -1,49 +1,45 @@
 package com.coodecool.pickyourspot.controller;
 
-import com.coodecool.pickyourspot.model.AppUser;
+import com.coodecool.pickyourspot.model.FoosballTable;
+import com.coodecool.pickyourspot.model.Reservation;
 import com.coodecool.pickyourspot.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
-@Controller
-@RequestMapping("/")
+import java.util.List;
+
+@RestController
+@RequestMapping("/api")
 public class TableController {
+    ProductService productService;
 
     @Autowired
-    private ProductService productService;
-    @GetMapping
-    public String index() {
-        return "index";
+    public TableController(ProductService productService) {
+        this.productService = productService;
     }
 
-    @GetMapping(value = "map")
-    public String displayMap() {
-        return "map";
+    @GetMapping("/table")
+    public List<FoosballTable> getAllTable(){
+        return productService.getAllTables();
     }
 
-    @GetMapping(value = "registration")
-    public String displayRegistration() {
-        return "registration";
+    @PostMapping("/table")
+    public void addTable(@RequestBody FoosballTable foosballTable){
+        productService.addNewTable(foosballTable);
     }
 
-    @PostMapping(value = "registration")
-    public String handleRegistration(@ModelAttribute("user") AppUser appUser) {
-        productService.addNewUser(appUser);
-        return "redirect:";
+    @PostMapping("/table/{tableId}/reservation")
+    public void addReservation(@RequestBody Reservation reservation, @PathVariable String tableId) throws IllegalAccessException {
+        productService.addReservation(tableId, reservation);
     }
 
-    @GetMapping(value= "login")
-    public String displayLogin() {
-        return "login";
+    @DeleteMapping("/table/{tableId}/reservation")
+    public void deleteReservation(@RequestBody Reservation reservation, @PathVariable String tableId){
+        productService.removeReservation(tableId, reservation);
     }
 
-    //TODO: login user create session
-    @PostMapping(value = "login")
-    public String handleLogin(@RequestBody AppUser appUser) {
-        return "redirect:";
+    @GetMapping("/table/free-tables/{dateTimeString}")
+    public List<FoosballTable> getFreeTables(@PathVariable String dateTimeString){
+        return productService.getFreeTables(dateTimeString);
     }
-
-
-
 }
